Move builder functions to a common helper
