+++ src/helpers/validateCommand.js
@@ -1,0 +1,28 @@
+import chalk from 'chalk';
+import has from 'has';
+
+import normalizeConfig from './normalizeConfig';
+import validateProjects from './validateProjects';
+import validateProject from './validateProject';
+
+export default function validateCommand(yargs) {
+  const config = normalizeConfig(yargs.argv);
+  const { project, projects, all } = config;
+  const allProjectNames = projects ? Object.keys(projects) : [];
+
+  if (all && allProjectNames.length <= 0) {
+    throw chalk.red('`--all` requires a non-empty “projects” config');
+  }
+  if (all && project) {
+    throw chalk.red('`--all` and `--project` are mutually exclusive');
+  }
+  if (project && !has(projects, project)) {
+    throw chalk.red(`Project "${project}" missing from “projects” config`);
+  }
+
+  if (projects) {
+    validateProjects(projects, allProjectNames, 'in the “projects” config');
+  } else {
+    validateProject(config);
+  }
+}
